feat: Allow setting JSON decoding strictness and parser in loads
The current default behavior of instantiating a `JSONDecoder` object is
to be strict. This means that the parser will raise an exception if it
encounters control characters in the input stream (like `\n` or `\t`).

Strings including control characters are common in LLM outputs, so this
change allows the user to set the strictness of the parser when calling
`loads`.

A simple way to use `loads` with a non-strict parser will be:

```python
from json_streamer import JsonParser, loads

mystream = stream_json()
ret = loads(mystream, parser=JsonParser(strict=False))
```